feat : migration to user login
Introduced iOS platform files including Xcode project, configuration, assets, and Swift/Objective-C sources. Added platform-specific authentication, token, and HTTP services, as well as responsive login/register pages for desktop and mobile. Updated main.dart and several widgets to support iOS integration.

+++ lib/widgets/mobile_widgets/main_page/total_assets_widget.dart
@@ -7,8 +7,8 @@ import 'package:gap/gap.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 Widget totalAssetsWidget(Map data){
-  int sbiBalance = data["variable_expense"] + data["savings"];
-  int iciciBalance = data["monthly_expense_left"].toDouble();
+  int sbiBalance = (data["variable_expense"] + data["savings"]).toInt();
+  int iciciBalance = data["monthly_expense_left"].toInt();
 
   int totalBalance = sbiBalance + iciciBalance;
